Add files via upload
This is for clearing the python shell while executing a terminal based programme. This also checks the operating system of the user for executing the command corresponding to it. 
Also, exception handling (try|except) cannot be used in this case because the error will be in the terminal and not the python shell, which is why the Python script won't recognise it and just keep on executing the script.